add feedbackController

+++ FreelancePlatform/src/main/java/freelanceplatform/data/FeedbackRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface FeedbackRepository extends CrudRepository<Feedback, Long> {
+public interface FeedbackRepository extends CrudRepository<Feedback, Integer> {
 
     List<Feedback> findAll();
